requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attr==0.3.1
attrs==24.2.0
azure-core==1.32.0
azure-storage-blob==12.24.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==5.0.1
boto==2.49.0
boto3==1.36.2
botocore==1.36.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
datamodel-code-generator==0.26.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
Django==4.2.18
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==2.4.0
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dlib==19.24.6
dnspython==2.7.0
docker-pycreds==0.4.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
easydict==1.13
email_validator==2.2.0
executing==2.1.0
expiringdict==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==33.3.1
fastapi==0.115.8
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.2.0
gast==0.6.0
genson==1.3.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==5.14.0
gradio_client==1.7.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h5py==3.12.1
hiddenlayer==0.3
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
imageio==2.36.0
importlib_metadata==8.5.0
inflect==5.6.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
json5==0.9.27
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.7
label-studio==1.15.0
label-studio-sdk==1.0.8
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
libclang==18.1.1
lockfile==0.12.2
loguru==0.7.2
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnx-simplifier==0.3.8
onnxoptimizer==0.3.13
onnxruntime==1.20.1
onnxruntime-gpu==1.20.1
onnxruntime-tools==1.7.0
openai==1.59.8
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.29.0
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.0.2
orjson==3.10.15
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
# pydensecrf @ file:///root/project/fgai-cv-basis/chapter07/pydensecrf
pydot==3.0.4
pydub==0.25.1
Pygments==2.18.0
PyMaxflow==1.3.2
pyparsing==3.2.0
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
python-multipart==0.0.20
pytorchcv==0.0.73
pytz==2022.7.1
PyYAML==6.0.2
pyzmq==26.2.0
redis==3.5.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rq==1.10.1
rsa==4.9
rstr==3.2.2
ruff==0.9.4
rules==3.4
s3transfer==0.11.1
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.3
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.2
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchnet==0.0.4
torchvision==0.19.1+cu121
torchviz==0.0.3
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.0.0
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
ultralytics==8.3.39
ultralytics-thop==2.0.12
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.34.0
visdom==0.2.4
wandb==0.18.7
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.40.0
widgetsnbextension==4.0.13
wrapt==1.17.0
xmljson==0.2.1
xmltodict==0.14.2
# -e git+https://github.com/Megvii-BaseDetection/YOLOX.git@d872c71bf63e1906ef7b7bb5a9d7a529c7a59e6a#egg=yolox
zipp==3.21.0
